% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dat.R
\name{convert_dat}
\alias{convert_dat}
\title{convert_dat}
\usage{
convert_dat(
  filenames,
  file_type = "rds",
  dir_in = NULL,
  dir_out = NULL,
  id = NULL,
  tz = NULL
)
}
\arguments{
\item{filenames}{vector; names of files (without path) to convert}

\item{file_type}{character string; 'rds' by default, any other will lead to conversion to csv}

\item{dir_in}{folder path of the .DAT files; if NULL, the working directory is used}

\item{dir_out}{folder path of the converted files; if NULL, the working directory is used}

\item{id}{character string; if not NULL, the string is used as prefix for the filenames of the converted files}

\item{tz}{time zone information; NULL (default) throws an error, as time is your most valuable information!
tz could be either a character string being a time zone code (e.g., "Europe/Paris") or a numeric value
indicating the number of hours to add/subtract to UTC time (e.g., -1 for times that would be UTC-1).}
}
\value{
files with converted data
}
\description{
This function convert the original .DAT files containing accelerometer or magnetometer data
into either rds (by default) or csv files.
}
