% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_samples.R
\name{extract_samples}
\alias{extract_samples}
\title{extract_samples}
\usage{
extract_samples(
  labels,
  path_audio,
  path_extracts = getwd(),
  ext_audio = ".wav"
)
}
\arguments{
\item{labels}{data.frame containing the information needed to extract and label samples}

\item{path_audio}{character string; path where the raw audio files are stored}

\item{path_extracts}{character string; path where to store the extracts}

\item{ext_audio}{character string; extension of the raw audio files}
}
\value{

}
\description{
This functions extract audio samples from wav files collected by the wehear logger, based on information
provided in label table.
}
