% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_inertial_pairs.R
\name{plot_inertial_pairs}
\alias{plot_inertial_pairs}
\title{plot_inertial_pairs}
\usage{
plot_inertial_pairs(df, title = "", radius = NULL, lim = NULL)
}
\arguments{
\item{df}{data.frame containing the data. Needs to have x,y, and z columns}

\item{title}{character string; title to be added to the plot}

\item{radius}{numeric; value of the radius of the circle to be drawn. If NULL (default), no circle is drawn}

\item{lim}{numeric; limit of the plot (in absolute value, the plot has an aspect ratio of 1);
If NULL (default), the choice is left to the ggplot function and all data are shown}
}
\value{

}
\description{
This function plot the values of inertial data across pairs of axes.
Useful for magnetometer calibration in particular.
}
