% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_inertial_ts.R
\name{plot_inertial_ts}
\alias{plot_inertial_ts}
\title{plot_inertial_ts}
\usage{
plot_inertial_ts(df, title = "")
}
\arguments{
\item{df}{data.frame containing the data. Needs to have x,y, and z columns}

\item{title}{character string; title to be added to the plot}
}
\value{

}
\description{
This function plots the time-series of inertial data
}
