% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{sync}
\alias{sync}
\title{sync}
\usage{
sync(port = "auto")
}
\arguments{
\item{port}{character string; either "auto" (default), which will use the last port (windows)
or the ttyUSB0 port (linux), or directly the name of a port (e.g., COM3)}
}
\value{

}
\description{
This functions serves to synchronise the wehear logger with the clock of a computer
}
