% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_pairs.R
\name{convert_to_pairs}
\alias{convert_to_pairs}
\title{convert_to_pairs}
\usage{
convert_to_pairs(df)
}
\arguments{
\item{df}{data.frame containing the data. Needs to have x,y, and z columns}
}
\value{
a data.frame with columns axis1, axis2 and pair (e.g. "x-y"). axis1 is the first of the two axes
of the pair, axis2 the second
}
\description{
This function reformats a data.frame containing inertial data so that values for pairs of axes
are more easily accessible. Mostly for internal use in other functions.
}
